+++ packages/react-scripts/template/src/App.jsx
@@ -6,22 +6,24 @@ import { BpkGridContainer, BpkGridRow, BpkGridColumn } from 'bpk-component-grid'
 
 import STYLES from './App.scss';
 
+const c = className => STYLES[className] || 'UNKNOWN';
+
 const App = () => (
-  <div className={STYLES.App}>
-    <header className={STYLES.App__header}>
+  <div className={c('App')}>
+    <header className={c('App__header')}>
       <BpkGridContainer>
         <BpkGridRow>
           <BpkGridColumn width={12}>
-            <BpkText tagName="h1" textStyle="xxl" className={STYLES.App__heading}>Welcome to React + Backpack</BpkText>
+            <BpkText tagName="h1" textStyle="xxl" className={c('App__heading')}>Welcome to React + Backpack</BpkText>
           </BpkGridColumn>
         </BpkGridRow>
       </BpkGridContainer>
     </header>
-    <main className={STYLES.App__main}>
+    <main className={c('App__main')}>
       <BpkGridContainer>
         <BpkGridRow>
           <BpkGridColumn width={12}>
-            <BpkText tagName="p" className={STYLES.App__text}>
+            <BpkText tagName="p" className={c('App__text')}>
               To get started, edit <BpkCode>src/App.jsx</BpkCode> and save to reload.
             </BpkText>
             <BpkButton onClick={() => alert('It works!')}>Click me</BpkButton>